+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+### Fixed
+- Fix a bug in `Libsvm::Node#==` when comparing with `nil`.
+
 ## [1.3.1] - 2015-03-21
 ### Fixed
 - Spelling and punctuation in documentation comments